+++ src/hosting/gitea_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"code.gitea.io/sdk/gitea"
+	"github.com/git-town/git-town/v9/src/cli"
 	"github.com/git-town/git-town/v9/src/config"
 	"github.com/git-town/git-town/v9/src/giturl"
 	"github.com/git-town/git-town/v9/src/hosting"
@@ -18,7 +19,7 @@ func TestNewGiteaConnector(t *testing.T) {
 			HostingService: config.HostingServiceGitea,
 			OriginURL:      giturl.Parse("git@custom-url.com:git-town/docs.git"),
 			APIToken:       "apiToken",
-			Log:            hosting.SilentPrinter{},
+			Log:            cli.SilentPrinter{},
 		})
 		assert.NoError(t, err)
 		wantConfig := hosting.CommonConfig{
@@ -35,7 +36,7 @@ func TestNewGiteaConnector(t *testing.T) {
 			HostingService: config.HostingServiceNone,
 			OriginURL:      giturl.Parse("git@github.com:git-town/git-town.git"),
 			APIToken:       "",
-			Log:            hosting.SilentPrinter{},
+			Log:            cli.SilentPrinter{},
 		})
 		assert.Nil(t, have)
 		assert.NoError(t, err)
@@ -47,7 +48,7 @@ func TestNewGiteaConnector(t *testing.T) {
 			HostingService: config.HostingServiceNone,
 			OriginURL:      originURL,
 			APIToken:       "",
-			Log:            hosting.SilentPrinter{},
+			Log:            cli.SilentPrinter{},
 		})
 		assert.Nil(t, have)
 		assert.NoError(t, err)
@@ -71,7 +72,7 @@ func TestGitea(t *testing.T) {
 			HostingService: config.HostingServiceGitea,
 			OriginURL:      giturl.Parse("git@gitea.com:git-town/docs.git"),
 			APIToken:       "",
-			Log:            hosting.SilentPrinter{},
+			Log:            cli.SilentPrinter{},
 		})
 		assert.Nil(t, err)
 		have, err := connector.NewProposalURL("feature", "parent")
@@ -83,7 +84,7 @@ func TestGitea(t *testing.T) {
 			HostingService: config.HostingServiceGitea,
 			OriginURL:      giturl.Parse("git@gitea.com:git-town/docs.git"),
 			APIToken:       "",
-			Log:            hosting.SilentPrinter{},
+			Log:            cli.SilentPrinter{},
 		})
 		assert.Nil(t, err)
 		have := connector.RepositoryURL()
